Move JOP to separate component
 - Move JOP Code editor and result to separate component
 - Add banner to scroll down to code editor
 - Update regex to allow for removal of whitespaces
 - Change texts
