+++ src/src/pages/FJerg/FJerg.scss
@@ -37,4 +37,12 @@
       }
     }
   }
+
+  .jop-code {
+    .code-result {
+      height: 100%;
+      display: flex;
+      align-items: center;
+    }
+  }
 }
